test(editor): Consolidate the N8nSelect exposed-API tests (no-changelog)
Absorbs #35113 and folds my own exposed-API block into it, so the contract is
pinned in one file with nothing asserted twice.

`Select.exposed.test.ts` keeps the union: the named editor-ui call sites
(`handleClose`, `$refs`), instance identity staying stable across a parent
re-render, the expose proxy yielding the raw instance rather than a Ref, clearing
on unmount, and the slot forwarding — plus the two assertions that were only in
`Select.test.ts`: that `focus`/`blur` actually delegate to the captured instance
rather than merely not throwing, and that `focusOnInput` moves focus to the inner
input. `Select.test.ts` goes back to rendering and interaction only.

Teardown is centralised in `afterEach` rather than unmounting inline. Every
wrapper here attaches to the document, and an attached wrapper left mounted
outlives the test environment — it reappears as an unhandled `document is not
defined` blamed on an unrelated file while every per-file result still reads
green. Inline unmounts only run when the assertions above them pass, so a
failing test could leak one.
